requirements.txt
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
defusedxml==0.6.0
Django==2.2
django-allauth==0.44.0
django-anymail==8.1
django-appconf==1.0.4
django-ckeditor==5.9.0
django-cors-headers==3.6.0
django-crispy-forms==1.10.0
django-debug-toolbar==3.2
django-extensions==3.1.0
django-filter==2.4.0
django-imagekit==4.0.2
django-js-asset==1.2.2
django-jsonfield==1.4.1
django-mptt==0.11.0
django-queryset-csv==1.1.0
django-rest-swagger==2.2.0
django-simple-history==3.0.0
django-tenants==3.1.0
djangorestframework==3.12.2
# djangorestframework-simplejwt==4.4.0
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.1.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
lml==0.1.0
mailjet-rest==1.3.4
Markdown==3.3.3
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
openpyxl==3.0.5
ordered-set==4.0.2
pilkit==2.0
# Pillow==6.2.1
# pkg-resources==0.0.0
psycopg2-binary==2.8.6
pycparser==2.20
pyexcel-io==0.6.4
pyexcel-xlsx==0.5.8
PyJWT==1.7.1
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.4
requests==2.23.0
requests-oauthlib==1.3.0
sentry-sdk==0.19.4
simplejson==3.17.2
six==1.15.0
#smartech-payment-api==0.1
sqlparse==0.4.1
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.25.11
xlwt==1.3.0
zipp==3.4.0
